+++ src/index.ts
@@ -3,13 +3,11 @@ export * from "./lib/Layout/AuthenticationLayout";
 
 // Panel sidebar layout
 export * from "./lib/Layout/PanelSideBarLayout/PanelSideBarLayout";
-export * from "./lib/Layout/PanelSideBarLayout/Context/PanelSideBarLayoutContext";
 export * from "./lib/Layout/PanelSideBarLayout/PanelSideBar/Context/PanelSideBarContext";
 export * from "./lib/Layout/PanelSideBarLayout/PanelSideBar/Definitions/PanelSideBarMenuItem";
 
 // Sidebar layout
 export * from "./lib/Layout/SideBarLayout/SideBarLayout";
-export * from "./lib/Layout/SideBarLayout/SideBarLayoutContext";
 export * from "./lib/SideBar/SideBarMenuContext";
 export * from "./lib/Paging/Paging";
 export * from "./lib/SideBar/ISideBarMenuItem";