+++ src/lib/Layout/PanelSideBarLayout/PanelSideBar/PanelSideBarItem.tsx
@@ -1,26 +1,24 @@
-import { IconProp } from "@fortawesome/fontawesome-svg-core";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import classNames from "classnames";
-import { ComponentType } from "react";
+import {ComponentType, useState} from "react";
 import { Collapse, NavItem } from "reactstrap";
 import { LinkRendererProps } from "src/lib/SideBar/SideBarMenuContext";
-import { PanelMenuItem } from "./Definitions/PanelSideBarMenuItem";
+import {PanelItem} from "src/lib/Layout/PanelSideBarLayout/PanelSideBar/Definitions/PanelSideBarMenuItem";
 
 export interface PanelSideBarItemProps {
-  item: PanelMenuItem<unknown>;
+  children: PanelItem<unknown>;
   LinkRenderer: ComponentType<LinkRendererProps>;
-  onClick?: (menuItem: PanelMenuItem<unknown>) => void;
+  onClick?: (menuItem: PanelItem<unknown>) => void;
   depth?: number;
   active?: boolean;
   toggledItemIds: string[];
 }
 
 const PanelSideBarItem = (props: PanelSideBarItemProps) => {
-  const { depth = 0, item, LinkRenderer, onClick, toggledItemIds = [] } = props;
+  const { depth = 0, children: item, LinkRenderer, onClick, toggledItemIds = [] } = props;
 
-  const hasChildren = !!item.children?.length;
-  const isOpen = toggledItemIds?.includes(item.id) || item.expanded;
-
+  const hasitem = !!item.children?.length;
+  const [isOpen, setIsOpen] = useState(toggledItemIds?.includes(item.id) || item.expanded);
   if (item.display === false) {
     return null;
   }
@@ -29,13 +27,13 @@ const PanelSideBarItem = (props: PanelSideBarItemProps) => {
     <>
       <NavItem
         onClick={() => onClick && onClick(item)}
-        className={classNames({ "menu-open": isOpen, active: item.active })}
+        className={classNames({ "menu-open": isOpen, active: item?.children ? item.children?.find(s => s.active) : item.active })}
         style={{ paddingLeft: depth ? `${depth + 1}rem` : undefined }}
       >
-        {hasChildren ? (
-          <div className={classNames({ dropend: !isOpen })}>
-            <a role="button" className={classNames("nav-link", { "dropdown-toggle": hasChildren })}>
-              {item.icon && <FontAwesomeIcon className="me-2" icon={item.icon as IconProp} />}
+        {hasitem ? (
+          <div>
+            <a role="button" className={classNames("nav-link", { "dropdown-toggle": hasitem })} onClick={() => setIsOpen(!isOpen)}>
+              {item.icon && <FontAwesomeIcon className="me-2" icon={item.icon} />}
               <div className="text-justify">{item.title}</div>
             </a>
           </div>
@@ -43,7 +41,7 @@ const PanelSideBarItem = (props: PanelSideBarItemProps) => {
           <>
             <LinkRenderer item={item}>
               <span className="nav-link">
-                {item.icon && <FontAwesomeIcon icon={item.icon as IconProp} className="me-2" />}
+                {item.icon && <FontAwesomeIcon icon={item.icon} className="me-2" />}
                 {item.title}
               </span>
             </LinkRenderer>
@@ -51,12 +49,12 @@ const PanelSideBarItem = (props: PanelSideBarItemProps) => {
         )}
       </NavItem>
 
-      {hasChildren && (
-        <Collapse isOpen={isOpen} navbar className={classNames("items-menu", { "mb-1": isOpen })}>
+      {hasitem && (
+        <Collapse isOpen={isOpen} navbar className={classNames("item-menu", { "mb-1": isOpen })}>
           {item.children?.map((childItem) => (
             <PanelSideBarItem
               key={childItem.id}
-              item={childItem}
+              children={childItem}
               LinkRenderer={LinkRenderer}
               onClick={() => onClick && onClick(childItem)}
               depth={depth + 1}